New abstract class SipsRequest
